Move handleRemoteClose and handleRemoteError to ViewController, so we have consistent behavior across views
